- Fixed some warnings. - mforms: added new node functions to manipulate existing child nodes in a treenode. Win + TUT implementation added too. - Win: mforms: switched off LoadOnDemand in TreeNodeAdv. This doesn't affect backend loading (e.g. in the schema tree) but avoids things like the expand icon to appear always for new or moved nodes, regardless whether they have child nodes or not. - grt: added ListRef move() call for List::move(). - Win: ANTLR parser: extended conditions to regenerate the parser files to the new simple parser. - trigger editor: added move up + down to the context menu for the trigger list.
[TESTS]
- Checked that treeviews (especially the schema tree) is not affected by switching off LoadOnDemand.
- Checked that triggers can be moved with the context menu, including to new groups, automatically changing their timing and event.

+++ backend/wbprivate/workbench/home_screen.cpp
@@ -80,9 +80,9 @@ static int image_width(cairo_surface_t* image)
   if (image != NULL)
   {
     if (mforms::Utilities::is_hidpi_icon(image) && mforms::App::get()->backing_scale_factor() > 1.0)
-      return cairo_image_surface_get_width(image) / mforms::App::get()->backing_scale_factor();
+      return (int)cairo_image_surface_get_width(image) / mforms::App::get()->backing_scale_factor();
     else
-      return cairo_image_surface_get_width(image);
+      return (int)cairo_image_surface_get_width(image);
   }
   return 0;
 }
@@ -3999,7 +3999,7 @@ public:
       for (ShortcutIterator iterator = _shortcuts.begin() + _page_start; iterator != _shortcuts.end();
         iterator++)
       {
-        double alpha = (yoffset + SHORTCUTS_ROW_HEIGHT) > height ? 0.25 : 1;
+        float alpha = (yoffset + SHORTCUTS_ROW_HEIGHT) > height ? 0.25 : 1;
 
         iterator->acc_bounds.pos.x = SHORTCUTS_LEFT_PADDING;
         iterator->acc_bounds.pos.y = yoffset;
